data_providers/trading_view.py
from typing import List
import requests

from data_structures.ohlc import OHLCData


class TradingViewProvider:
    """TradingView data provider (using their unofficial API)"""
    
    def __init__(self):
        self.session = requests.Session()
        self._setup_session()
    
    def _setup_session(self):
        """Setup session with TradingView"""
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
            'Referer': 'https://www.tradingview.com/',
            'Origin': 'https://www.tradingview.com'
        })
    
    def get_ohlc_data(self, symbol: str, timeframe: str, count: int = 100) -> List[OHLCData]:
        """Fetch OHLC data from TradingView (simplified implementation)"""
        
        # This is a simplified implementation
        # For production, you'd want to use TradingView's official API or websocket
        
        print(f"TradingView provider: Would fetch {symbol} {timeframe} data (not implemented in demo)")
        return []

